# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AuthKit API
  version: 0.1.0-rc.1
  description: Static integration contract. Runtime documentation endpoints are intentionally disabled.
servers:
  - url: https://auth.example.com
security:
  - bearerAuth: []
paths:
...
  /api/v1/auth/login:
    post:
      security: []
      requestBody: {$ref: '#/components/requestBodies/Login'}
      responses: {'200': {description: Successful response}, '400': {description: Invalid request}, '429': {$ref: '#/components/responses/RateLimited'}, '503': {description: Dependency unavailable}}
  /api/v1/auth/mfa/verify-login:
    post:
      security: []
      requestBody: {$ref: '#/components/requestBodies/MfaLoginVerification'}
      responses: {'200': {description: Successful response}, '400': {description: Invalid request}, '429': {$ref: '#/components/responses/RateLimited'}, '503': {description: Dependency unavailable}}
  /api/v1/auth/passkeys/options:
    post:
      security: []
      requestBody: {$ref: '#/components/requestBodies/PasskeyAssertionOptions'}
      responses: {'200': {description: Successful response}, '400': {description: Invalid request}, '429': {$ref: '#/components/responses/RateLimited'}, '503': {description: Dependency unavailable}}
  /api/v1/auth/passkeys/verify:
    post:
      security: []
      requestBody: {$ref: '#/components/requestBodies/PasskeyAssertionFinish'}
      responses: {'200': {description: Successful response}, '400': {description: Invalid request}, '429': {$ref: '#/components/responses/RateLimited'}, '503': {description: Dependency unavailable}}
  /api/v1/auth/refresh:
    post:
      security: [{refreshCookie: [], csrfHeader: []}]
      responses: {'200': {$ref: '#/components/responses/Success'}, '401': {$ref: '#/components/responses/Error'}}
  /api/v1/auth/logout:
    post:
      security: [{refreshCookie: [], csrfHeader: []}]
      responses: {'200': {$ref: '#/components/responses/Success'}, '401': {$ref: '#/components/responses/Error'}}
  /api/v1/auth/logout-all:
    post:
      requestBody: {$ref: '#/components/requestBodies/MfaOptionalVerification'}
      responses: {'200': {description: Successful response}, '401': {description: Unauthorized}, '403': {description: Forbidden}}
  /api/v1/auth/register:
    post:
      security: []
      requestBody: {$ref: '#/components/requestBodies/Register'}
      responses: {'201': {description: Created in non-stealth mode}, '202': {description: Accepted without account-enumeration disclosure}, '400': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}, '429': {$ref: '#/components/responses/RateLimited'}, '503': {description: Dependency unavailable}}
  /api/v1/auth/email-confirmation/confirm:
    post:
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/EmailConfirmationConfirmRequest'}
      responses: {'200': {description: Successful response}, '400': {description: Invalid request}, '429': {$ref: '#/components/responses/RateLimited'}, '503': {description: Dependency unavailable}}
  /api/v1/auth/email-confirmation/resend:
    post:
      security: []
      requestBody: {$ref: '#/components/requestBodies/EmailConfirmationResend'}
      responses: {'200': {$ref: '#/components/responses/Success'}, '400': {$ref: '#/components/responses/Error'}, '429': {$ref: '#/components/responses/RateLimited'}, '503': {description: Dependency unavailable}}
  /api/v1/auth/email-change/confirm:
    post:
      security: []
      requestBody: {$ref: '#/components/requestBodies/EmailChangeConfirm'}
      responses: {'200': {$ref: '#/components/responses/Success'}, '400': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}, '429': {$ref: '#/components/responses/RateLimited'}, '503': {description: Dependency unavailable}}
  /api/v1/auth/password-recovery/request:
    post:
      security: []
      requestBody: {$ref: '#/components/requestBodies/PasswordRecovery'}
      responses: {'200': {$ref: '#/components/responses/Success'}, '400': {$ref: '#/components/responses/Error'}, '429': {$ref: '#/components/responses/RateLimited'}, '503': {description: Dependency unavailable}}
  /api/v1/auth/password-recovery/reset:
    post:
      security: []
      parameters:
        - {name: email, in: query, required: true, schema: {type: string, format: email, maxLength: 255}}
      requestBody: {$ref: '#/components/requestBodies/PasswordReset'}
      responses: {'200': {$ref: '#/components/responses/Success'}, '400': {$ref: '#/components/responses/Error'}, '429': {$ref: '#/components/responses/RateLimited'}, '503': {description: Dependency unavailable}}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: First-party AuthKit access token with `token_use=first_party_access`, configured API audience, non-revoked key, and active server-side session. OAuth access tokens and ID tokens are rejected.
    oauthAccessBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OAuth access token with `token_use=oauth_access`, OAuth client audience, `client_id`, and server-side scope checks. ID tokens and first-party access tokens are rejected.
    refreshCookie:
      type: apiKey
      in: cookie
      name: Refresh-Token
      description: HttpOnly refresh cookie scoped to AuthKit refresh/logout routes.
    csrfHeader:
      type: apiKey
      in: header
      name: X-XSRF-TOKEN
      description: CSRF header that must echo the readable CSRF cookie on refresh and logout.
    oauthClientBasic:
      type: http
      scheme: basic
      description: OAuth client authentication for token, introspection, and revocation endpoints. Public clients may use approved body credentials where supported.
    workerToken:
      type: apiKey
      in: header
      name: X-Worker-Token
      description: Rotatable internal credential accepted only from explicitly trusted network peers.
...
  requestBodies:
    Login: &requiredLoginBody {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/LoginRequest'}}}}
    MfaLoginVerification: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/MfaLoginVerificationRequest'}}}}
    PasskeyAssertionOptions: {required: false, content: {application/json: {schema: {$ref: '#/components/schemas/PasskeyAssertionOptionsRequest'}}}}
    PasskeyAssertionFinish: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/PasskeyAssertionFinishRequest'}}}}
    Register: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/RegisterRequest'}}}}
    EmailConfirmationResend: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/EmailConfirmationResendRequest'}}}}
    EmailChangeConfirm: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/EmailChangeConfirmRequest'}}}}
    PasswordRecovery: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/PasswordRecoveryRequest'}}}}
    PasswordReset: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/PasswordResetRequest'}}}}
...
    MfaOptionalVerification: {required: false, content: {application/json: {schema: {$ref: '#/components/schemas/MfaOptionalVerificationRequest'}}}}
...
  headers:
    XRequestId:
      required: true
      schema: {type: string, format: uuid}
      description: Request correlation identifier; a valid inbound X-Request-ID is preserved, otherwise AuthKit generates one.
  responses:
    Success:
      description: Successful response
      headers: {X-Request-ID: {$ref: '#/components/headers/XRequestId'}}
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiResponse'}}}
    Error:
      description: Opaque error response
      headers: {X-Request-ID: {$ref: '#/components/headers/XRequestId'}}
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiResponse'}}}
...
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After: {required: true, schema: {type: integer, minimum: 1}, description: Seconds before retrying}
        X-Request-ID: {$ref: '#/components/headers/XRequestId'}
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiResponse'}}}
...
  schemas:
    LoginRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email: {type: string, format: email, maxLength: 255, example: person@example.com}
        password: {type: string, minLength: 1, maxLength: 128, writeOnly: true, example: correct-horse-battery-staple}
    MfaLoginVerificationRequest:
      type: object
      additionalProperties: false
      required: [mfaToken, code]
      properties:
        mfaToken: {type: string, minLength: 1, maxLength: 512, writeOnly: true}
        code: {type: string, minLength: 6, maxLength: 32, writeOnly: true}
    PasskeyAssertionOptionsRequest:
      type: object
      additionalProperties: false
      properties: {email: {type: string, format: email, maxLength: 255}}
    PasskeyAssertionFinishRequest:
      type: object
      additionalProperties: false
      required: [challengeId, credentialJson]
      properties:
        challengeId: {type: string, format: uuid}
        credentialJson: {type: string, minLength: 1, maxLength: 20000, writeOnly: true}
    RegisterRequest:
      type: object
      additionalProperties: false
      required: [email, password, termsAccepted, privacyPolicyAccepted]
      properties:
        email: {type: string, format: email, maxLength: 255}
        password: {type: string, minLength: 12, maxLength: 128, writeOnly: true}
        termsAccepted: {type: boolean, enum: [true]}
        privacyPolicyAccepted: {type: boolean, enum: [true]}
    EmailConfirmationResendRequest:
      type: object
      additionalProperties: false
      required: [email]
      properties: {email: {type: string, format: email, maxLength: 255}}
    EmailChangeConfirmRequest:
      type: object
      additionalProperties: false
      required: [token]
      properties: {token: {type: string, minLength: 1, maxLength: 512, writeOnly: true}}
    PasswordRecoveryRequest:
      type: object
      additionalProperties: false
      required: [email]
      properties: {email: {type: string, format: email, maxLength: 255}}
    PasswordResetRequest:
      type: object
      additionalProperties: false
      required: [token, newPassword]
      properties:
        token: {type: string, minLength: 1, maxLength: 512, writeOnly: true}
        newPassword: {type: string, minLength: 12, maxLength: 128, writeOnly: true}
...
    ApiResponse:
      type: object
      additionalProperties: false
      required: [timestamp, requestId]
      properties:
        data: {}
        errors: {type: array, items: {type: string}, nullable: true}
        timestamp: {type: string, format: date-time}
        code: {type: string, pattern: '^[a-z0-9_]+$', nullable: true, description: Stable machine-readable error code}
        requestId: {type: string, minLength: 1, maxLength: 128}
...
    MfaOptionalVerificationRequest:
      type: object
      additionalProperties: false
      properties:
        currentPassword: {type: string, maxLength: 128, nullable: true, writeOnly: true}
        code: {type: string, minLength: 6, maxLength: 32, nullable: true, writeOnly: true}
...
    EmailConfirmationConfirmRequest:
      type: object
      additionalProperties: false
      required: [token]
      properties:
        token:
          type: string
          minLength: 32
          maxLength: 128
          description: One-time activation token read by the frontend from the URL fragment and submitted in the JSON body.